Refactors cookies, navigation, and timeouts into json_wire_spec.
