+++ lib/sync_core.dart
@@ -28,10 +28,11 @@ export 'package:webdriver/src/sync/command_processor.dart';
 export 'package:webdriver/src/sync/common.dart';
 export 'package:webdriver/src/sync/exception.dart';
 export 'package:webdriver/src/sync/json_wire_spec/keyboard.dart';
+export 'package:webdriver/src/sync/common_spec/cookies.dart';
 export 'package:webdriver/src/sync/json_wire_spec/logs.dart';
 export 'package:webdriver/src/sync/json_wire_spec/mouse.dart';
-export 'package:webdriver/src/sync/navigation.dart';
-export 'package:webdriver/src/sync/options.dart';
+export 'package:webdriver/src/sync/common_spec/navigation.dart';
+export 'package:webdriver/src/sync/timeouts.dart';
 export 'package:webdriver/src/sync/target_locator.dart';
 export 'package:webdriver/src/sync/web_driver.dart';
 export 'package:webdriver/src/sync/web_element.dart';
